requirements.txt
streamlit== 1.44.1
tensorflow== 2.19.0
numpy== 2.1.0
pandas== 2.2.2
plotly== 6.3.0
Pillow== 10.4.0